Refactor wallet endpoints into its own module
