+++ marlowe-dashboard-server/app/PSGenerator.hs
@@ -15,22 +15,24 @@ module PSGenerator
   )
 where
 
-import           Control.Applicative                ((<|>))
-import           Control.Lens                       (set, (&))
-import           Data.Monoid                        ()
-import           Data.Proxy                         (Proxy (Proxy))
-import qualified Data.Text.Encoding                 as T ()
-import qualified Data.Text.IO                       as T ()
-import           Language.PureScript.Bridge         (BridgePart, Language (Haskell), SumType, argonaut, buildBridge,
-                                                     typeName, writePSTypes, (^==))
-import           Language.PureScript.Bridge.PSTypes (psNumber, psString)
-import           Language.PureScript.Bridge.SumType (equal, genericShow, mkSumType, order)
-import           Marlowe.Run.Webserver.API          (HTTPAPI)
-import           Marlowe.Run.Webserver.Types        (RestoreError, RestorePostData)
-import           Marlowe.Run.Webserver.WebSocket    (StreamToClient, StreamToServer)
+import           Control.Applicative                                   ((<|>))
+import           Control.Lens                                          (set, (&))
+import           Data.Monoid                                           ()
+import           Data.Proxy                                            (Proxy (Proxy))
+import qualified Data.Text.Encoding                                    as T ()
+import qualified Data.Text.IO                                          as T ()
+import           Language.PureScript.Bridge                            (BridgePart, Language (Haskell), SumType,
+                                                                        argonaut, buildBridge, typeName, writePSTypes,
+                                                                        (^==))
+import           Language.PureScript.Bridge.PSTypes                    (psNumber, psString)
+import           Language.PureScript.Bridge.SumType                    (equal, genericShow, mkSumType, order)
+import           Marlowe.Run.Webserver.API                             (HTTPAPI)
+import           Marlowe.Run.Webserver.Wallet.CentralizedTestnet.Types (RestoreError, RestorePostData)
+import           Marlowe.Run.Webserver.WebSocket                       (StreamToClient, StreamToServer)
 import qualified PSGenerator.Common
-import           Servant.PureScript                 (HasBridge, Settings, apiModuleName, defaultBridge, defaultSettings,
-                                                     languageBridge, writeAPIModuleWithSettings)
+import           Servant.PureScript                                    (HasBridge, Settings, apiModuleName,
+                                                                        defaultBridge, defaultSettings, languageBridge,
+                                                                        writeAPIModuleWithSettings)
 
 doubleBridge :: BridgePart
 doubleBridge = typeName ^== "Double" >> return psNumber
