+++ marlowe-dashboard-server/marlowe-dashboard-server.cabal
@@ -22,8 +22,13 @@ library
         Marlowe.Run.Webserver.API
         Marlowe.Run.Webserver.WebSocket
         Marlowe.Run.Webserver.Types
+        Marlowe.Run.Webserver.Wallet.CentralizedTestnet.Types
+        Marlowe.Run.Webserver.Wallet.Client
     other-modules:
         Paths_marlowe_dashboard_server
+        Marlowe.Run.Webserver.Wallet.API
+        Marlowe.Run.Webserver.Wallet.CentralizedTestnet.API
+        Marlowe.Run.Webserver.Wallet.CentralizedTestnet.Server
     hs-source-dirs:
         src
     default-language: Haskell2010
