+++ marlowe-dashboard-server/src/Marlowe/Run/Webserver/API.hs
@@ -1,17 +1,12 @@
-{-# LANGUAGE DataKinds          #-}
-{-# LANGUAGE DeriveAnyClass     #-}
-{-# LANGUAGE DeriveGeneric      #-}
-{-# LANGUAGE DerivingStrategies #-}
-{-# LANGUAGE TypeOperators      #-}
+{-# LANGUAGE DataKinds     #-}
+{-# LANGUAGE TypeOperators #-}
 
 module Marlowe.Run.Webserver.API where
 
 import           Cardano.Prelude
-import           Marlowe.Run.Webserver.Types (RestoreError, RestorePostData)
-import           Servant.API                 (Get, JSON, PlainText, Post, Raw, ReqBody, (:<|>), (:>))
-import           Servant.API.WebSocket       (WebSocketPending)
--- FIXME: I don't like to use a Mock type here, but we'd need to publish some changes upstream to the PAB to fix this
-import           Cardano.Wallet.Mock.Types   (WalletInfo)
+import qualified Marlowe.Run.Webserver.Wallet.API as Wallet
+import           Servant.API                      (Get, JSON, PlainText, Raw, (:<|>), (:>))
+import           Servant.API.WebSocket            (WebSocketPending)
 
 type API = WebSocketAPI
     :<|> HTTPAPI
@@ -19,9 +14,7 @@ type API = WebSocketAPI
 
 type HTTPAPI = "api" :>
     ("version" :> Get '[PlainText, JSON] Text
-    :<|> "wallet" :>
-        ("restore" :> ReqBody '[ JSON] RestorePostData :> Post '[JSON] (Either RestoreError WalletInfo)
-        )
+    :<|> "wallet" :> Wallet.API
     )
 
 type WebSocketAPI = "ws" :> WebSocketPending
