+++ marlowe-dashboard-server/src/Marlowe/Run/Webserver/Types.hs
@@ -1,30 +1,7 @@
-{-# LANGUAGE DeriveAnyClass     #-}
-{-# LANGUAGE DeriveGeneric      #-}
-{-# LANGUAGE DerivingStrategies #-}
-
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Marlowe.Run.Webserver.Types where
 
-import           Cardano.Prelude
-import           Data.Aeson.Types (FromJSON, ToJSON)
-import           Servant.Client   (ClientEnv)
-
-
-data RestorePostData =
-    RestorePostData
-        { getMnemonicPhrase :: [Text]
-        , getPassphrase     :: Text
-        , getWalletName     :: Text
-        }
-    deriving stock (Eq, Generic, Show)
-    deriving anyclass (FromJSON, ToJSON)
-
-data RestoreError =
-    InvalidMnemonic
-    | RestoreWalletError
-    | FetchPubKeyHashError
-    deriving stock (Eq, Generic, Show)
-    deriving anyclass (ToJSON)
+import           Servant.Client (ClientEnv)
 
 type Env = ClientEnv
